Fixed VS building problems.
git-svn-id: https://svnsrv.germanlloyd.org/svn/ExtFEMIO/trunk@184 389f8db8-dd28-11e4-b689-8b400cb78fb4

+++ include/fem/cards.h
@@ -247,7 +247,7 @@ Description:
             const long *TYPE, const long *SUBTYPE,
             const ::std::deque<::std::string> *CONT) :
             TYPE(*TYPE), SUBTYPE(*SUBTYPE), CONT(*CONT) {
-            NRECS = CONT->size();
+            NRECS = static_cast<long>(CONT->size());
             NBYTE = 0;
             for (auto &p : *CONT)
               NBYTE = (NBYTE < (long)p.size()) ? (long)p.size() : NBYTE;
@@ -346,7 +346,7 @@ DATE      0.10000000E+01  0.00000000E+00  0.40000000E+01  0.72000000E+02
             const long *TYPE, const long *SUBTYPE,
             const ::std::deque<::std::string> *CONT) :
             TYPE(*TYPE), SUBTYPE(*SUBTYPE), CONT(*CONT) {
-            NRECS = CONT->size();
+            NRECS = static_cast<long>(CONT->size());
             NBYTE = 0;
             for (auto &p : *CONT)
               NBYTE = (NBYTE < (long)p.size()) ? (long)p.size() : NBYTE;
